requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
watchdog==3.0.0
requests==2.31.0